requirements.txt
iniconfig==2.0.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
pyreadr==0.5.2
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
tqdm==4.67.0
tzdata==2024.2
